requirements_BACKUP_1205.txt
pygame-ce
pygame_gui
bcrypt
